backend/pyproject.toml
[project]
name = "begautos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach>=6.2.0",
    "dj-database-url>=2.3.0",
    "django>=5.1.4",
    "django-ckeditor>=6.7.1",
    "django-colorfield>=0.12.0",
    "django-cors-headers>=4.6.0",
    "django-countries>=7.6.1",
    "django-csp>=3.8",
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "django-ipware>=7.0.1",
    "django-jazzmin>=3.0.0",
    "django-phonenumber-field>=8.0.0",
    "django-tinymce>=4.1.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-yasg>=1.21.8",
    "email-verifier>=1.0.0",
    "faker>=33.1.0",
    "fontawesomefree>=6.5.1",
    "geoip2>=4.8.0",
    "gunicorn>=23.0.0",
    "markdown>=3.7",
    "phonenumbers>=8.13.52",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "user-agents>=2.2.0",
]
